D3ASIM-3553-market-agent: Rename area_agents module to market_agents
