+++ src/gsy_e/models/area/event_dispatcher.py
@@ -34,13 +34,13 @@ from gsy_e.models.area.redis_dispatcher.market_notify_event_subscriber import (
     MarketNotifyEventSubscriber)
 from gsy_e.models.market import MarketBase
 from gsy_e.models.market.market_structures import AvailableMarketTypes
-from gsy_e.models.strategy.area_agents.balancing_agent import BalancingAgent
-from gsy_e.models.strategy.area_agents.future_agent import FutureAgent
-from gsy_e.models.strategy.area_agents.one_sided_agent import OneSidedAgent
-from gsy_e.models.strategy.area_agents.one_sided_alternative_pricing_agent import (
+from gsy_e.models.strategy.market_agents.balancing_agent import BalancingAgent
+from gsy_e.models.strategy.market_agents.future_agent import FutureAgent
+from gsy_e.models.strategy.market_agents.one_sided_agent import OneSidedAgent
+from gsy_e.models.strategy.market_agents.one_sided_alternative_pricing_agent import (
     OneSidedAlternativePricingAgent)
-from gsy_e.models.strategy.area_agents.settlement_agent import SettlementAgent
-from gsy_e.models.strategy.area_agents.two_sided_agent import TwoSidedAgent
+from gsy_e.models.strategy.market_agents.settlement_agent import SettlementAgent
+from gsy_e.models.strategy.market_agents.two_sided_agent import TwoSidedAgent
 
 if TYPE_CHECKING:
     from gsy_e.models.area import Area
