+++ src/gsy_e/models/strategy/market_agents/balancing_agent.py
@@ -18,8 +18,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from numpy.random import random
 from gsy_e.constants import FLOATING_POINT_TOLERANCE
 from gsy_e.gsy_e_core.util import make_ba_name, make_ma_name
-from gsy_e.models.strategy.area_agents.one_sided_agent import OneSidedAgent
-from gsy_e.models.strategy.area_agents.one_sided_engine import BalancingEngine
+from gsy_e.models.strategy.market_agents.one_sided_agent import OneSidedAgent
+from gsy_e.models.strategy.market_agents.one_sided_engine import BalancingEngine
 from gsy_framework.constants_limits import ConstSettings
