+++ src/gsy_e/models/strategy/market_agents/future_agent.py
@@ -20,8 +20,8 @@ from typing import TYPE_CHECKING
 from gsy_framework.constants_limits import ConstSettings
 
 from gsy_e.gsy_e_core.util import make_fa_name
-from gsy_e.models.strategy.area_agents.future_engine import FutureEngine
-from gsy_e.models.strategy.area_agents.two_sided_agent import TwoSidedAgent
+from gsy_e.models.strategy.market_agents.future_engine import FutureEngine
+from gsy_e.models.strategy.market_agents.two_sided_agent import TwoSidedAgent
 
 if TYPE_CHECKING:
     from gsy_e.models.area import Area
