+++ src/gsy_e/models/strategy/market_agents/settlement_agent.py
@@ -18,8 +18,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from gsy_framework.constants_limits import ConstSettings
 
 from gsy_e.gsy_e_core.util import make_sa_name
-from gsy_e.models.strategy.area_agents.two_sided_agent import TwoSidedAgent
-from gsy_e.models.strategy.area_agents.two_sided_engine import TwoSidedEngine
+from gsy_e.models.strategy.market_agents.two_sided_agent import TwoSidedAgent
+from gsy_e.models.strategy.market_agents.two_sided_engine import TwoSidedEngine
 
 
 class SettlementAgent(TwoSidedAgent):
