+++ src/gsy_e/models/strategy/market_agents/two_sided_agent.py
@@ -18,8 +18,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from gsy_framework.constants_limits import ConstSettings
 from numpy.random import random
 
-from gsy_e.models.strategy.area_agents.one_sided_agent import OneSidedAgent
-from gsy_e.models.strategy.area_agents.two_sided_engine import TwoSidedEngine
+from gsy_e.models.strategy.market_agents.one_sided_agent import OneSidedAgent
+from gsy_e.models.strategy.market_agents.two_sided_engine import TwoSidedEngine
 
 
 class TwoSidedAgent(OneSidedAgent):
